[✨Feat] #39 - 대기 상세 뷰 API 연동
